More cleanup.

+++ Headers/diis.h
@@ -35,9 +35,9 @@
  
 template <typename T> class diis {
 public:
-    diis(int max_Min, int N_in);
-    int max_M;
-    int N;
+    diis(size_t max_Min, size_t N_in);
+    size_t max_M;
+    size_t N;
     double lambda;
     double eps = 1.0e-11;           // epsilon added to B_ii for stability
     bool cleared = false;
